+++ plotnine/guides/guide_axis.py
@@ -1,0 +1,12 @@
+from .guide import guide
+
+
+class guide_axis(guide):
+    """
+    Axis
+    """
+    # parameter
+    available_aes = {'x', 'y'}
+
+    def draw(self):
+        pass